Moved kill/unkill button into CashOutPage, made logic for clicking apps in the selectApp activity:
