--- app/src/main/res/layout/item_app1.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/container"
-    android:layout_width="match_parent"
-    android:layout_height="180dp"
-    android:foregroundTint="#352424"
-    android:orientation="vertical"
-    android:padding="4dp">
-
-    <ImageView
-        android:id="@+id/appIcon"
-        android:layout_width="80dp"
-        android:layout_height="120dp"
-        android:layout_gravity="center_horizontal"
-        tools:srcCompat="@tools:sample/avatars" />
-
-    <TextView
-        android:id="@+id/appName"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="center"
-        tools:text="App Name" />
-</LinearLayout>
